Migrate 'visitable' test
