+++ tests/helpers/properties.js
@@ -1,7 +1,10 @@
 import { AcceptanceAdapter, moduleForAcceptance, testForAcceptance } from './properties/acceptance-adapter';
 import { IntegrationAdapter, moduleForIntegration, testForIntegration } from './properties/integration-adapter';
 
-export function moduleForProperty(name, cb) {
+export function moduleForProperty(name, cbOrOptions, cb) {
+  let options = cb ? cbOrOptions : {};
+  cb = cb || cbOrOptions;
+
   // Generate acceptance tests
   let acceptanceAdapter = new AcceptanceAdapter();
   moduleForAcceptance(`Acceptance mode | Property | ${name}`, {
@@ -10,6 +13,10 @@ export function moduleForProperty(name, cb) {
     }
   });
   cb(testForAcceptance, acceptanceAdapter);
+
+  if (options.acceptanceOnly) {
+    return;
+  }
 
   // Generate integration tests
   let integrationAdapter = new IntegrationAdapter();
